Issue #1
Transform CSV file data into a JSON file. This is done in a new folder (utils) in the src folder. When running the cnvertCSVtoJSON.js a new output.js file will be created. This will then contaijn the initial input data for the table in the webapplication and the database. It is not yet linked to the webapp.